define useSpinner hook
