+++ packages/experimental/Spinner/src/SpinnerTokens.android.ts
@@ -8,6 +8,4 @@ import type { SpinnerTokens } from './Spinner.types';
 export const defaultSpinnerTokens: TokenSettings<SpinnerTokens, Theme> = () =>
   ({
     trackColor: Appearance.getColorScheme() === 'light' ? globalTokens.color.grey56 : globalTokens.color.grey72,
-    lineThickness: 'medium',
-    size: 'medium',
   } as SpinnerTokens);
